Move model filtering and union construction to overture-schema-core
Extracts filter_models() and resolve_types() into core/discovery.py.
Union-building helpers (create_union_type_from_models, _discriminated_union,
_can_discriminate, _type_literal) go to a new core/union.py module.

ModelDict and UnionType type aliases live alongside the functions that
produce them (discovery.py and union.py respectively).

Both CLI and the upcoming validation package import these from core,
avoiding a cross-dependency.
