+++ packages/overture-schema-core/src/overture/schema/core/__init__.py
@@ -1,15 +1,30 @@
 from . import cartography, names, scoping, sources
+from .discovery import (
+    ModelDict,
+    ModelKey,
+    discover_models,
+    filter_models,
+    resolve_types,
+)
 from .models import OvertureFeature, ThemeT, TypeT
 from .scoping import Scope, scoped
+from .union import UnionType, create_union_type_from_models
 
 __all__ = [
     "cartography",
+    "create_union_type_from_models",
+    "discover_models",
+    "filter_models",
+    "ModelDict",
+    "ModelKey",
     "names",
     "OvertureFeature",
+    "resolve_types",
     "Scope",
     "scoped",
     "scoping",
     "sources",
     "ThemeT",
     "TypeT",
+    "UnionType",
 ]
